rename to riverpod_analyzer_utils
