+++ examples/counter/pubspec_overrides.yaml
@@ -1,4 +1,4 @@
-# melos_managed_dependency_overrides: flutter_riverpod,hooks_riverpod,riverpod,riverpod_lint
+# melos_managed_dependency_overrides: flutter_riverpod,hooks_riverpod,riverpod,riverpod_lint,riverpod_analyzer_utils
 dependency_overrides:
   flutter_riverpod:
     path: ../../packages/flutter_riverpod
@@ -6,5 +6,7 @@ dependency_overrides:
     path: ../../packages/hooks_riverpod
   riverpod:
     path: ../../packages/riverpod
+  riverpod_analyzer_utils:
+    path: ../../packages/riverpod_analyzer_utils
   riverpod_lint:
     path: ../../packages/riverpod_lint
