--- packages/riverpod_analysis/lib/riverpod_analysis.dart
@@ -1,7 +1,0 @@
-export 'src/utils.dart';
-export 'src/type_checker.dart';
-export 'src/analyzer_utils.dart';
-export 'src/riverpod_types.dart';
-export 'src/extensions.dart';
-export 'src/dependencies.dart';
-export 'src/visitors.dart';
