+++ packages/riverpod_analyzer_utils/lib/src/dependencies.dart
