+++ packages/riverpod_analyzer_utils/lib/src/extensions.dart
