+++ packages/riverpod_analyzer_utils/lib/src/type_checker.dart
