+++ packages/riverpod_analyzer_utils/lib/src/utils.dart
