+++ packages/riverpod_analyzer_utils/pubspec.yaml
@@ -1,4 +1,4 @@
-name: riverpod_analysis
+name: riverpod_analyzer_utils
 description: A sample command-line application.
 version: 1.0.0
 # repository: https://github.com/my_org/my_repo
