+++ packages/riverpod_generator/pubspec.yaml
@@ -12,12 +12,12 @@ dependencies:
   collection: ^1.15.0
   crypto: ^3.0.2
   meta: ^1.7.0
+  riverpod_analyzer_utils: any
   riverpod_annotation: ^1.0.0
   source_gen: ^1.2.0
-  riverpod_analysis: any
 
 dev_dependencies:
   build_runner: ^2.1.7
+  build_test: any
   riverpod: ^2.0.0
   test: ^1.21.0
-  build_test: any
