+++ packages/riverpod_lint/bin/custom_lint.dart
@@ -5,9 +5,8 @@ import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer_plugin/protocol/protocol_generated.dart';
 import 'package:analyzer_plugin/utilities/change_builder/change_builder_core.dart';
-import 'package:collection/collection.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
-import 'package:riverpod_analysis/riverpod_analysis.dart';
+import 'package:riverpod_analyzer_utils/riverpod_analyzer_utils.dart';
 
 void main(List<String> args, SendPort sendPort) {
   startPlugin(sendPort, RiverpodLint());
