+++ packages/riverpod_lint/pubspec.yaml
@@ -14,9 +14,9 @@ dependencies:
   meta: ^1.7.0
   path: ^1.8.1
   riverpod: ^2.0.0
+  riverpod_analyzer_utils: any
   source_span: ^1.8.0
   yaml: ^3.1.1
-  riverpod_analysis: any
 
 dev_dependencies:
   hotreloader: ^3.0.5
