+++ packages/riverpod_lint/pubspec_overrides.yaml
@@ -2,5 +2,5 @@
 dependency_overrides:
   riverpod:
     path: ../riverpod
-  riverpod_analysis:
-    path: ../riverpod_analysis
+  riverpod_analyzer_utils:
+    path: ../riverpod_analyzer_utils
